requirements.txt
python-telegram-bot[job-queue]
hummingbot-api-client==1.2.8
python-dotenv
pytest
pre-commit
black
isort
PyYAML
aiohttp
watchfiles
kaleido
plotly
pandas
geckoterminal-py
mcp
pytest-asyncio
